+++ weblate/urls.py
@@ -732,6 +732,11 @@ real_patterns = [
         name="js-translate",
     ),
     path(
+        "js/dismiss-automatically-translated/<int:unit_id>/",
+        weblate.trans.views.js.dismiss_automatically_translated,
+        name="js-dismiss-automatically-translated",
+    ),
+    path(
         "js/memory/<int:unit_id>/",
         weblate.machinery.views.memory,
         name="js-memory",